Integration layer for MassCraxx propositions 2
AutoCook File split
tooltip typo
variable scope
x4 Font handling
