+++ AutoCook/media/lua/shared/Translate/EN/UI_EN.txt
@@ -8,9 +8,9 @@ UI_EN = {
     UI_AutoCookUseRotten = "Use Rotten Ingredients",
     UI_AutoCookUseRottenTooltip = "Enables using rotten ingredients for AutoCook, if character is able to.",
     UI_AutoCookMaxDuplicate = "Duplicate ingredients:",
-    UI_AutoCookMaxDuplicateTooltip = "How many ingredients of the same type are allowed in your cooking.",
+    UI_AutoCookMaxDuplicateTooltip = "How many ingredients of the same type are allowed.",
     UI_AutoCookMaxSpices = "Spices used:",
-    UI_AutoCookMaxSpicesTooltip = "How many spices are used in your cooking. If smart spices is active, this is overridden when close to under/overweight.",
+    UI_AutoCookMaxSpicesTooltip = "How many spices are used. <LINE>If smart spices is active, this is overridden when close to under/overweight.",
     UI_AutoCookSmartSpices = "Smart Spices",
     UI_AutoCookSmartSpicesTooltip = "Overrides spices used when close to under/overweight.",
     UI_AutoCookFreshness = "Freshness",
